Revert "Muiz branch"
