+++ front-end/package.json
@@ -6,14 +6,12 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "^5.3.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
-    "react-router-dom": "^6.22.0",
+    "react-router-dom": "^6.21.3",
     "react-scripts": "5.0.1",
-    "reactjs-popup": "^2.0.6",
     "reactstrap": "^9.2.2",
     "web-vitals": "^2.1.4"
   },
