--- front-end/src/CSS/index.css
@@ -1,33 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Kotta+One&family=Protest+Riot&display=swap');
-*{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-
-}
-
-body {
-  margin: 0;
-  font-family: 'Kotta One', serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-html, body {
-  font-family: 'Protest_Riot', fallback-font, sans-serif;
-  height: 100%;
-}
-
-#root {
-  height: 100%;
-}
