--- front-end/src/Container/Home/Home.css
@@ -1,5 +1,0 @@
-.main-container{
-    position: absolute;
-    z-index: 1;
-    top: 25vh;
-}
