+++ front-end/src/Containers/Dailies/Dailies.js
