+++ front-end/src/Containers/Journal/Journal.js
