+++ front-end/src/Containers/Signup/signup.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+function Signup(){
+    return(
+        <div className="main-container">sign up test</div>
+    );
+}
+export default Signup;
