+++ front-end/src/Css/index.css
@@ -1,28 +1,4 @@
 /* All Font Settings Should be Here */
-@import url('https://fonts.googleapis.com/css2?family=Kotta+One&family=Protest+Riot&display=swap');
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-
-}
-
-body {
-    margin: 0;
-    font-family: 'Kotta One', serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-
-}
-
-code {
-    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
 
 @font-face {
     font-family: 'Protest_Riot';
