+++ front-end/src/Css/signin.css
@@ -1,0 +1,49 @@
+:root {
+    --log-in-form-bg-color: #c4d3ad; 
+  }
+
+  .log-in-form {
+    min-height: 60vh;
+    margin:auto;
+  }
+
+.form-content {
+    margin:auto;
+    margin-top: 5%;
+    padding: 2%;
+    width: 40%;
+    background-color: var(--log-in-form-bg-color);
+    align-items: center;
+    text-align: center;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    border: 1px solid #ddd;
+}
+
+.log-in-form .form-content .formgroup {
+    padding-top:1%;
+    padding-bottom: 1%;
+    margin-right: 0;
+    margin-left: 0;
+    width: 100%;
+}
+
+.log-in-form .form-content Input {
+    width: 70%;
+}
+
+.submit-btn {
+    background-color: #548f68;
+    margin-bottom: 3%;
+    width: 30%;
+    font-size: large;
+}
+
+
+.lg-font {
+    font-size: xx-large;
+}
+
+span {
+    padding-top: 2%;
+}
